correccion de organizacion de chess y board
