+++ main/chess.py
@@ -9,6 +9,42 @@ class Chess:
     def turn(self):
         return self.__turn__
     
+    # Captura la pieza en la posicion indicada
+    def capture_piece(self, from_row, from_col, to_row, to_col):
+        piece = self.__board__.get_piece(from_row, from_col)
+        if piece is None:
+            return None  # No piece at the start position
+
+        target_piece = self.__board__.get_piece(to_row, to_col)
+        if target_piece and target_piece.get_color() != piece.get_color():
+            self.__board__.remove_piece(to_row, to_col)  # Remove the captured piece
+            self.__board__.__positions__[from_row][from_col] = None
+            self.__board__.__positions__[to_row][to_col] = piece
+            piece.__board__.set_position(to_row, to_col)
+            return True
+        
+    # Mueve la pieza y detecta si no hay pieza en esa posicion y si el destino esta ocupado
+    def move_piece(self, from_row, from_col, to_row, to_col):
+        piece = self.__board__.get_piece(from_row, from_col)
+         
+        target_piece = self.__board__.get_piece(to_row, to_col)
+        if target_piece and target_piece.get_color() == piece.get_color():
+            raise InvalidMoveSameColor("La posicion tiene una pieza del mismo color")
+
+        #if self.is_valid_move(to_row, to_col, self):
+        if self.capture_piece(from_row, from_col, to_row, to_col):
+            return True  # Pieza en el destino es del enemigo y se captura
+        
+        # Mueve la pieza
+        #if self.is_valid_move(to_row, to_col, self):
+            self.__positions__[from_row][from_col] = None
+            self.__positions__[to_row][to_col] = piece
+            piece.set_position(to_row, to_col)
+            return True
+        #else:
+        #    raise InvalidMove("Movimiento invalido")
+
+
     def move(self, from_row, from_col, to_row, to_col):
         move_piece_color = self.__board__.get_color(from_row, from_col)
         final_piece_color = self.__board__.get_color(to_row, to_col)
@@ -22,13 +58,15 @@ class Chess:
         if move_piece_color == final_piece_color:
             raise InvalidMoveSameColor("La posicion tiene una pieza del mismo color")
         
-        result = self.__board__.move_piece(from_row, from_col, to_row, to_col)
+        result = self.move_piece(from_row, from_col, to_row, to_col)
         if result:
             self.change_turn()
             return True
         self.change_turn()
         return False
     
+
+
     def change_turn(self):
         self.__turn__ = "BLACK" if self.__turn__ == "WHITE" else "WHITE"
